+++ src/main/java/com/alumnione/ecommerce/model/order/repository/OrderRepository.java
@@ -1,6 +1,6 @@
-package com.alumnione.ecommerce.model.orders.repository;
+package com.alumnione.ecommerce.model.order.repository;
 
-import com.alumnione.ecommerce.model.orders.entity.Order;
+import com.alumnione.ecommerce.model.order.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
