+++ src/main/java/com/alumnione/ecommerce/model/payment/repository/PaymentRepository.java
@@ -1,6 +1,6 @@
-package com.alumnione.ecommerce.model.payments.repository;
+package com.alumnione.ecommerce.model.payment.repository;
 
-import com.alumnione.ecommerce.model.payments.entity.Payment;
+import com.alumnione.ecommerce.model.payment.entity.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PaymentRepository extends JpaRepository<Payment, Long> {
